stack/app/members/requirements.txt
pymonad
pynamodb
attrdict
web3
pyjwt

lambda_decorators
six